+++ RegistrationWindow/RegistrationWindow.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
