+++ RegistrationWindow/RegistrationWindow/Assets.xcassets/AccentColor.colorset/Contents.json
