+++ RegistrationWindow/RegistrationWindow/Assets.xcassets/AppIcon.appiconset/Contents.json
