+++ RegistrationWindow/RegistrationWindow/Assets.xcassets/Contents.json
