+++ RegistrationWindow/RegistrationWindow/Base.lproj/LaunchScreen.storyboard
