+++ RegistrationWindow/RegistrationWindow/CellModel.swift
