+++ RegistrationWindow/RegistrationWindow/Info.plist
