+++ RegistrationWindow/RegistrationWindow/View/HeaderView.swift
